+++ Changelog.md
@@ -3,6 +3,8 @@ Group The-real-jim-parry-for-real-this-time
 Jason Sekhon(Pilot), Brayden Traas(Co-pilot), Eric Lin(Cook), Morris Arroyo(Stewardess)
 
 -------------------------
+* Updated flight model unit test to cover half of flight model's properties
+
 * Updated airplane model unit test to cover all of airplane model's  properties
 
 * added ./test/tmp/* and ./tmp/ to gitignore